+++ sdmx-dl-provider-base/src/main/java/sdmxdl/provider/ConnectionSupport.java
@@ -4,6 +4,7 @@ import lombok.NonNull;
 import sdmxdl.*;
 
 import java.io.IOException;
+import java.io.UncheckedIOException;
 import java.util.stream.Stream;
 
 @lombok.experimental.UtilityClass
@@ -21,6 +22,8 @@ public class ConnectionSupport {
     public static @NonNull DataSet getDataSetFromStream(@NonNull CatalogRef catalog, @NonNull FlowRef flowRef, @NonNull Query query, @NonNull Connection connection) throws IOException, IllegalArgumentException {
         try (Stream<Series> stream = connection.getDataStream(catalog, flowRef, query)) {
             return stream.collect(DataSet.toDataSet(flowRef, query));
+        } catch (UncheckedIOException ex) {
+            throw ex.getCause();
         }
     }
 }